# xpand.client/openapi/model/update-planet-dto.ts
/**
 * EXPAND.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanetStatus } from './planet-status';


export interface UpdatePlanetDto { 
    planetId: string;
    status: PlanetStatus;
    description?: string | null;
    teamId?: string | null;
}