# xpand.client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EXPAND.Server",
    "version": "v1"
  },
  "paths": {
    "/api/planets": {
      "get": {
        "tags": [
          "Planet"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Planet"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPlanetDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPlanetDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddPlanetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/planets/{id}": {
      "get": {
        "tags": [
          "Planet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Planet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/planets/{id})": {
      "patch": {
        "tags": [
          "Planet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanetDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanetDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/team/{id}/captain": {
      "get": {
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CaptainDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptainDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptainDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/team/{id}/robots": {
      "get": {
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RobotDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RobotDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RobotDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddPlanetDto": {
        "required": [
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/PlanetStatus"
          },
          "teamId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CaptainDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PlanetDto": {
        "required": [
          "id",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/PlanetStatus"
          },
          "teamId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlanetStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "ToDo",
          "EnRoute",
          "OK",
          "Not_OK"
        ]
      },
      "RobotDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdatePlanetDto": {
        "required": [
          "planetId",
          "status"
        ],
        "type": "object",
        "properties": {
          "planetId": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PlanetStatus"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "teamId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}